add cypress and testcafe tutorials

+++ packages/eyes-cypress/tutorial/run.sh
@@ -1,12 +1,12 @@
 #!/bin/bash
 set -e
-rm -r ./package
+[ -d "./package" ] && rm -r ./package
+echo $(ls)
 mkdir package
 cd ..
 yarn pack
 package=$(find applitools*.tgz)
 mv "$package" ./tutorial/package/"$package"
 cd ./tutorial
-docker-compose build
-docker-compose run wdio4_selenium_basic
-docker-compose run wdio4_selenium_ultrafastgrid
+docker build -t tutorial_cypress .
+docker run -e APPLITOOLS_API_KEY tutorial_cypress
