+++ packages/eyes-testcafe/tutorial/run.sh
@@ -1,12 +1,11 @@
 #!/bin/bash
 set -e
-rm -r ./package
+[ -d "./package" ] && rm -r ./package
 mkdir package
 cd ..
 yarn pack
 package=$(find applitools*.tgz)
 mv "$package" ./tutorial/package/"$package"
 cd ./tutorial
-docker-compose build
-docker-compose run wdio4_selenium_basic
-docker-compose run wdio4_selenium_ultrafastgrid
+docker build -t tutorial_testcafe .
+docker run -e APPLITOOLS_API_KEY tutorial_testcafe
